Implement combo condition for cart details template
- Restructuring the session based Cart() class to use a dictionary called 'items'
to keep track of all the items, tours in this case.
- Implementing a method inside the Cart() class to determine if the contents in
a given time of the cart are a combo, if they are, return True and also set the
id of the combo into the session.
- Placeholding in the cart details a section to make obvious that a combo induced
discount is happening and which one it is. Removed most of the previous "sale"
functionalities from the cart details template.
- Making the django admins Category/Combo creation also save the many2many
field "tours", previously it only saved after creation.
- Adding the tour type and duration to the tour cards in the search templates (
missed the combo template).
- Making the combo tags green in the tour details template to differentiate them
from common categories.